requirements.txt
paho_mqtt==1.6.1
pyserial==3.5
pyubx2==1.2.27
